refactor(axelar): split test files by destination chain, to work with mocha  parallelization
