+++ modules/axelar/test/gmp/xrpl-evm/call-contract.test.ts
@@ -48,18 +48,16 @@ describeOrSkip(
         const pollingOpts = config.xrplEvmChain as PollingOptions;
 
         before(async () => {
-            const { urls: destUrls, axelarExecutableExampleAddress, interchainTokenExecutableExampleAddress } = xrplEvmChain;
-
-            destAxExecAddress = axelarExecutableExampleAddress;
-            destIntTokenExecAddress = interchainTokenExecutableExampleAddress;
+            destAxExecAddress = xrplEvmChain.axelarExecutableExampleAddress;
+            destIntTokenExecAddress = xrplEvmChain.interchainTokenExecutableExampleAddress;
 
             xrplClient = new Client(xrplChain.urls.ws);
-            evmJsonProvider = new ethers.JsonRpcProvider(destUrls.rpc);
+            evmJsonProvider = new ethers.JsonRpcProvider(xrplEvmChain.urls.rpc);
 
             xrplChainProvider = new XrplProvider(xrplClient);
 
-            xrplChainWallet = Wallet.fromSeed(xrplChain.account.privateKey);
-            evmChainWallet = new ethers.Wallet(xrplEvmChain.account.privateKey, evmJsonProvider);
+            xrplChainWallet = Wallet.fromSeed(xrplChain.accounts.privateKeys[2]);
+            evmChainWallet = new ethers.Wallet(xrplEvmChain.accounts.privateKeys[0], evmJsonProvider);
 
             xrplChainSigner = new XrplSigner(xrplChainWallet, xrplChainProvider);
